Example tests and associated tidying up of test artifacts.
